Updated Angular v19.2.1

+++ db/schema.rb
@@ -27,7 +27,7 @@ ActiveRecord::Schema[8.0].define(version: 2022_02_20_134729) do
     t.string "content_type"
     t.text "metadata"
     t.string "service_name", null: false
-    t.bigint "byte_size", null: false
+    t.integer "byte_size", null: false
     t.string "checksum"
     t.datetime "created_at", null: false
     t.index ["key"], name: "index_active_storage_blobs_on_key", unique: true
